Extract drawing button attributes to editor extension method
