+++ Assets/EasyButtons/Editor/ButtonEditor.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Linq;
-using System.Reflection;
-using UnityEngine;
-using UnityEditor;
+﻿using UnityEditor;
 
 namespace EasyButtons
 {
@@ -15,34 +11,7 @@ namespace EasyButtons
     {
         public override void OnInspectorGUI()
         {
-            // Loop through all methods with no parameters
-            var methods = target.GetType()
-                .GetMethods(BindingFlags.Instance | BindingFlags.Public | BindingFlags.NonPublic)
-                .Where(m => m.GetParameters().Length == 0);
-            foreach (var method in methods)
-            {
-                // Get the ButtonAttribute on the method (if any)
-                var ba = (ButtonAttribute)Attribute.GetCustomAttribute(method, typeof(ButtonAttribute));
-
-                if (ba != null)
-                {
-                    // Determine whether the button should be enabled based on its mode
-                    GUI.enabled = ba.Mode == ButtonMode.AlwaysEnabled
-                        || (EditorApplication.isPlaying ? ba.Mode == ButtonMode.EnabledInPlayMode : ba.Mode == ButtonMode.DisabledInPlayMode);
-
-                    // Draw a button which invokes the method
-                    var buttonName = String.IsNullOrEmpty(ba.Name) ? ObjectNames.NicifyVariableName(method.Name) : ba.Name;
-                    if (GUILayout.Button(buttonName))
-                    {
-                        foreach (var t in targets)
-                        {
-                            method.Invoke(t, null);
-                        }
-                    }
-
-                    GUI.enabled = true;
-                }
-            }
+            this.DrawEasyButtons();
 
             // Draw the rest of the inspector as usual
             DrawDefaultInspector();
